Refactor file imports to kebab casing in app.tsx and related components (#8)
