--- src/components/DonationDialog/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./DonationDialog";
