--- src/components/ResultBox/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./ResultBox";
