+++ src/components/donation-dialog/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./donation-dialog";
