+++ src/components/result-box/components/value-table-cell.tsx
