+++ src/components/result-box/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./result-box";
